Create main activity and add manifest file
Main activity consists of a fragment container, that shows
one of two main fragments of the application and a bottom
navigation, that allows changing current fragment.
Activity remembers last opened fragment and shows it after
screen unlock.
Add android manifest with screen orientation locked to a
portrait view, since dialogs are pain in the ass.