Display a placeholder message instead of an empty recycler view
In the ActivitiesStartedFragment use EmptiableRecyclerView instead of
a usual RecyclerView.
Make ActivitiesStartEventsAdapter compatible with a new
EmptiableRecyclerView by making the first one inherit ListAdapter
instead of a usual Adapter.
Create a message repository, with messages to display, when there are
no activities, started today. Those messages are displayed by the
EmptiableRecyclerView.
Add set of a fallback messages to display.