Fix EmpitableRecyclerView swipe state hanging bug
Hide actual recycler view instead of removing it with visibility "gone",
so that notification from it's adapter about removal of the last item
properly updates the state of the item, being "swiped out" of the list.
Otherwise, recycler view not re-render itself according to
the notification and when the next activity start event will be created,
for a moment it will be red, because item's swipe state has not been
properly changed.